Support nested hardware virtualization for VirtualBox, too.
This requires VirtualBox 6.1 or later. In my experimenting, I found that
the Proxmox VE node hosting the KVM VM sometimes hung when it demanded
access to all four cores. Reducing this to 1 didn't impact the VM's
performance (for the purposes of this example) and greatly improved the
reliability of the provisioning scripts.